Integrated Firebase across the project, added new features

+++ lib/Pages/favPage.dart
@@ -1,4 +1,5 @@
-import 'package:ecomerce_app/Custom%20Widgets/FavItem.dart';
+import 'package:ecomerce_app/Custom%20Widgets/favItem.dart';
+import 'package:ecomerce_app/Pages/itemPage.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -9,36 +10,54 @@ class FavPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final listener = context.watch<itemProvider>();
-    final provider = context.read<itemProvider>();
+    final listener = context.watch<ItemProvider>();
+    final provider = context.read<ItemProvider>();
 
     return Scaffold(
-        appBar: AppBar(title: Text("Favourite items",style:TextStyle(fontSize: 25,fontWeight: FontWeight.w700,),),
-        automaticallyImplyLeading: false,
+        appBar: AppBar(
+          title: const Text(
+            "Favourite items",
+            style: TextStyle(
+              fontSize: 25,
+              fontWeight: FontWeight.w700,
+            ),
+          ),
+          automaticallyImplyLeading: false,
         ),
-    body: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
-      provider.favItemSet.toList().isNotEmpty
-          ? Expanded(
-              child: ListView.builder(
-                  itemCount: provider.favItemSet.length,
-                  itemBuilder: (context, index) {
-                    return FavItem(
-                      itemPrice: provider.favItemSet.toList()[index]["price"],
-                      imgPath:
-                          "assets/home/${provider.favItemSet.toList()[index]["name"].toString().toLowerCase()}.png",
-                      itemName:
-                          '${provider.favItemSet.toList()[index]["name"]}',
-                      deleteItem: () {
-                        listener.deleteFavItem(index);
-                      },
-                    );
-                  }),
-            )
-          : Center(
-              child: Text(
-              "No items have been added to favourites",
-              style: TextStyle(fontSize: 1),
-            ))
-    ]));
+        body: StreamBuilder(
+            stream: provider.favStream,
+            builder: (context, snapshot) {
+              provider.getFavStream();
+              listener.getFavDocs(snapshot);
+
+
+              if (snapshot.connectionState == ConnectionState.waiting) {
+                return const Center(child: CircularProgressIndicator());
+              } else if ( provider.favDocs.isEmpty) {
+                return const Center(child: Text("Not items in Favourites"));
+              } else {
+
+
+
+                return ListView.builder(
+                    itemCount: provider.favDocs.length,
+                    itemBuilder: (context, index) {
+                      return FavItem(
+                        onTap: (){
+                          Navigator.push(context, MaterialPageRoute(builder: (context){
+                            return Itempage(item: provider.favDocs[index]);
+                          }));
+                        },
+                        color: provider.selectedColor,
+                        itemPrice: provider.favDocs[index]["price"],
+                        imgPath: provider.favDocs[index]["imageUrl"],
+                        itemName: '${provider.favDocs[index]["name"]}',
+                        deleteItem: () {
+                          listener.deleteFav(index);
+                        },
+                      );
+                    });
+              }
+            }));
   }
 }
